prodinfra/prod-dashbord/Backend/requirements.txt


django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-allauth
django-rest-auth[with_social]
django-environ
django-debug-toolbar
django-extensions
psycopg2-binary
gunicorn
djangorestframework-jsonapi
daphne
channels
django-redis
django-filter
django-axes
django-secure
pytest-django
factory_boy
django-storages
django-crispy-forms
dj_database_url
Pillow
mysqlclient
boto3
cryptography
python-jose[cryptography]
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
redis
celery
mysqlclient

drf-spectacular  
django-guardian  
whitenoise  
django-health-check 
django-ratelimit
 dj-database-url
 cryptography